move dataset to init
